backend/app/requirements.txt
flask
flask-cors
PyJWT
flask-jwt-extended
flask-sqlalchemy
psycopg2-binary
openai==0.28
pydub
moviepy
numpy
PyPDF2
python-docx
transformers
celery
redis
requests
beautifulsoup4
tldextract
azure-storage-blob
Authlib